+++ projects/mtg/src/MTGAbility.cpp
@@ -469,6 +469,13 @@ int AbilityFactory::parseCastRestrictions(MTGCardInstance * card, Player * playe
 				return 0;
 		}
 
+        check = restriction[i].find("canuntap");
+        if(check != string::npos)
+        {
+            if(card->frozen >= 1 || card->basicAbilities[(int)Constants::DOESNOTUNTAP] || !card->isTapped())
+                return 0;
+        }
+
         check = restriction[i].find("raid");
         if(check != string::npos)
         {
@@ -2226,6 +2233,15 @@ MTGAbility * AbilityFactory::parseMagicLine(string s, int id, Spell * spell, MTG
         return a;
     }
 
+    //gain control until source is untapped or leaves battlefield
+    found = s.find("shackle");
+    if (found != string::npos)
+    {
+        MTGAbility * a = NEW AShackleWrapper(observer, id, card, target);
+        a->oneShot = 1;
+        return a;
+    }
+
    //momentary blink
     found = s.find("(blink)");
     if (found != string::npos)
@@ -2802,8 +2818,7 @@ MTGAbility * AbilityFactory::parseMagicLine(string s, int id, Spell * spell, MTG
     vector<string> splitCascade = parseBetween(s, "cascade:", " ", false);
     if (splitCascade.size())
     {
-        Targetable * t = spell ? spell->getNextTarget() : NULL;
-        MTGAbility * a = NEW AACascade(observer, id, card, t , splitCascade[1], NULL, who);
+        MTGAbility * a = NEW AACascade(observer, id, card, target, splitCascade[1], NULL);
         a->oneShot = 1;
         return a;
     }